requirements.txt
numpy
opencv-python
fastapi
pydantic
keras
tensorflow
torch
torchvision
matplotlib
scikit-image
Pillow